+++ package.json
@@ -4,7 +4,8 @@
   "description": "Tiny web service for parsing yarn.lock files",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js"
+    "start": "node index.js",
+    "test": "mocha"
   },
   "engines": {
     "node": "9.8.0"
@@ -29,5 +30,11 @@
     "@yarnpkg/lockfile": "^1.0.0",
     "body-parser": "^1.15.2",
     "express": "^4.14.0"
+  },
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "mocha": "^5.2.0",
+    "request": "^2.87.0",
+    "test": "^0.6.0"
   }
 }
